SMSEditor initialization updates
updated the constructor logic based on results of unit tests.  If the constructor is called without passing in a symbol, it will load the default symbol.  Otherwise, it will  use the passed in symbol.  This should work for both editing default symbol as well as a specific graphics symbol.

Also created full suite of unit tests for SMSEditor